# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Readify API Documentation",
        "description": "API documentation for Readify application",
        "contact": {
            "email": "muhammadzelvan263@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Readify API Server"
        }
    ],
    "paths": {
...
        "/book-shelves": {
            "get": {
                "tags": [
                    "Book Shelves"
                ],
                "summary": "Get list of book shelves",
                "operationId": "2dca346dce3b0ae2aa86ba021930aec5",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by code or location",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of book shelves retrieved successfully"
                    }
                }
            },
            "post": {
                "tags": [
                    "Book Shelves"
                ],
                "summary": "Create a new book shelf",
                "operationId": "597bee85885a3d837e0b732f6e6b4d75",
                "parameters": [
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Book shelf code",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Book shelf location",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "capacity",
                        "in": "query",
                        "description": "Book shelf capacity",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Book shelf description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Book shelf created successfully"
                    }
                }
            }
        },
        "/book-shelves/{id}": {
            "get": {
                "tags": [
                    "Book Shelves"
                ],
                "summary": "Get a book shelf by ID",
                "operationId": "6627b6f674c7dcb6c3783adce14f83f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book shelf ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book shelf retrieved successfully"
                    }
                }
            },
            "put": {
                "tags": [
                    "Book Shelves"
                ],
                "summary": "Update a book shelf by ID",
                "operationId": "b69f1e50cba0fb0b8888febc5d2cc1fd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book shelf ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "Book shelf code",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "location",
                        "in": "query",
                        "description": "Book shelf location",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "capacity",
                        "in": "query",
                        "description": "Book shelf capacity",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "Book shelf description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book shelf updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Book Shelves"
                ],
                "summary": "Delete a book shelf by ID",
                "operationId": "fc28c72a9f338b9008ede61ebe72f501",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Book shelf ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Book shelf deleted successfully"
                    }
                }
            }
        },
...
    },
...
}